refactor: update imports to use /index.ts
