+++ src/components/savings/index.js
@@ -1,4 +1,2 @@
 export { default as APYPill } from './APYPill';
-export { default as SavingsSheetEmptyState } from './SavingsSheetEmptyState';
-export { default as SavingsPredictionStepper } from './SavingsPredictionStepper';
-export { default as SavingsSheetHeader } from './SavingsSheetHeader';
+export { default as SavingsListHeader } from './SavingsListHeader';
