+++ src/components/send/SendAssetList.js
@@ -19,7 +19,7 @@ import {
   SendSavingsCoinRow,
 } from '../coin-row';
 import { Centered } from '../layout';
-import SavingsListHeader from '../savings/SavingsListHeader';
+import { SavingsListHeader } from '../savings';
 import TokenFamilyHeader from '../token-family/TokenFamilyHeader';
 import { layoutEasingAnimation } from '@cardstack/utils';
 import { ImgixImage } from '@rainbow-me/images';
